Encryption Update
I replaced "user communications" with "user information" in the first two indicators and added a new indicator on encryption of information at rest. In the research guidance (glossary and extended description of what each indicator measures) we will need to explain that "user information" includes the sorts of data Ross mentioned about such as command and control messaging, software and hardware updates, etc.