Update authentication.yaml
Add two indicators relevant to the authentication system of IoT devices and mobile applications. The company should only obtain the user's password hashes for storage in order to reduce the attack surface of the entire authentication system. The hash algorithm should be secure. For example, the MD5 hash may be decoded via rainbow table lookup.